+++ NEWS.rst
@@ -1,4 +1,4 @@
-Changes in 3.16 (released ??/??/2010)
+Changes in 3.16 (released ??/??/2011)
 -------------------------------------
 
 *	:mod:`ll.sisyphus` logging of exceptions and tracebacks should be more robust
@@ -13,6 +13,10 @@ Changes in 3.16 (released ??/??/2010)
 *	Fixed a bug in the command line argument handling of ``dtd2xsc.py``.
 
 *	``dtd2xsc.py`` can now handle undefined entities.
+
+*	The UL4 function ``json`` and ``type`` can now handle all instances of
+	:class:`collections.Mapping` and :class:`collections.Sequence` not just
+	:class:`tuple`, :class:`list` and :class:`dict`.
 
 
 Changes in 3.15.3 (released 11/26/2010)